feat: moving tax id formatter to utils
