+++ src/service/customer/CustomerService.tsx
@@ -3,18 +3,9 @@ import LocalEnvironment from "../../config/LocalEnvironment";
 import { CustomerResponse } from "../../dto/customer/CustomerResponse";
 import { Severity, Variant } from "../../components/toaster/ToasterProvider";
 import { CustomerRegisterFormData } from "../../dto/customer/CustomerRegisterFormData";
+import formatUtils from "../../utils/format/FormatUtils";
 
 class CustomerService {
-
-    formatCNPJ = (cnpj: string): string => {
-        const onlyDigits = cnpj.replace(/\D/g, '');
-
-        if (cnpj !== onlyDigits) {
-            return onlyDigits;
-        }
-
-        return onlyDigits.replace(/^(\d{2})(\d{3})(\d{3})(\d{4})(\d{2})$/, "$1.$2.$3/$4-$5");
-    };
 
     customerUrl = LocalEnvironment.API_URL + '/customer';
 
@@ -27,7 +18,7 @@ class CustomerService {
             );
             return response.data.map((item: any) => ({
                 name: item.name,
-                taxId: this.formatCNPJ(item.taxId)
+                taxId: formatUtils.formatCNPJ(item.taxId)
             }));
         } catch {
             toaster('Ocorreu um erro, tente novamente mais tarde.', 5000, 'error', 'filled');
@@ -42,7 +33,7 @@ class CustomerService {
             this.customerUrl,
             {
                 'name': customerRegisterFormData.name,
-                'taxId': this.formatCNPJ(customerRegisterFormData.taxId)
+                'taxId': formatUtils.formatCNPJ(customerRegisterFormData.taxId)
             },
             {
                 headers: {
